Changes to ceilometer plugin: 1. Rename ceilometer driver from opencontrail to contrail 2. Change contrail driver to support only the floating IP meters
